Fixes for scons builds
1) Remove oic-utilities as a requirement

2) Fail the build if one step fails

Setting this option in the bash script will cause the
script to terminate if any of the commands returns a
with non-zero value.  This is necessary for the build
server to understand when a build is failed.  Without this
flag, the script will exit with the status of the last command,
which in this script is "echo" so it will be 0, regardless if
a scons build step failed.

Ideally we would try each step and then report the failures
at the end, set -e will do the job until that can be developed.

3) Conditionaly enable android build

4) Correct a build issue in occoap.c with typecasting

+++ resource/csdk/occoap/src/occoap.c
@@ -131,7 +131,7 @@ static void HandleCoAPRequests(struct coap_context_t *ctx,
     coap_block_t rcvdBlock2;
     memset(&rcvdBlock1, COAP_BLOCK_FILL_VALUE, sizeof(coap_block_t));
     memset(&rcvdBlock2, COAP_BLOCK_FILL_VALUE, sizeof(coap_block_t));
-    size_t rcvdSize1 = 0;
+    uint16_t rcvdSize1 = 0;
     coap_pdu_t * rcvdPdu = rcvdRequest->pdu;
     coap_pdu_t * sendPdu = NULL;
     coap_send_flags_t sendFlag;
